unit testing, move of store into its own directory
