+++ jest.config.js
@@ -1,3 +1,5 @@
+//  process.env.VUE_CLI_BABEL_TARGET_NODE = true
+//  process.env.VUE_CLI_BABEL_TRANSPILE_MODULES = true
 module.exports = {
   collectCoverage: true,
   collectCoverageFrom: ["**/*.{js,vue}", "!**/node_modules/**","!**/dist/**"],
