+++ package.json
@@ -5,7 +5,8 @@
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
-    "test:unit": "vue-cli-service test:unit"
+    "test:unit": "vue-cli-service test:unit",
+    "test:debug": "node --inspect node_modules/.bin/jest --runInBand"
   },
   "dependencies": {
     "axios": "^0.19.0",
@@ -16,7 +17,7 @@
     "d3": "^5.9.2",
     "flush-promises": "^1.0.2",
     "jquery": "^3.4.1",
-    "vue": "^2.6.6",
+    "vue": "^2.6.10",
     "vue-chartjs": "^3.4.2",
     "vue-instant": "^1.0.2",
     "vue-router": "^3.0.1",
