+++ src/main.js
@@ -5,7 +5,7 @@ import 'bootstrap/dist/css/bootstrap.css'
 import 'bootstrap-vue/dist/bootstrap-vue.css'
 import App from './App.vue'
 import router from './router'
-import store from './store'
+import store from './store/store'
 import './assets/css/main.css'
 import axios from 'axios'
