+++ src/store/mutation-types.js
@@ -1,0 +1,5 @@
+//pattern desribed in https://vuex.vuejs.org/guide/mutations.html
+export const AUTH_REQUEST = "AUTH_REQUEST";
+export const AUTH_SUCCESS = "AUTH_SUCCESS";
+export const AUTH_ERROR = "AUTH_ERROR";
+export const LOGOUT = "LOGOUT";
