+++ src/store/store.js
@@ -1,6 +1,7 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
 import axios from 'axios'
+import {AUTH_REQUEST, AUTH_SUCCESS, AUTH_ERROR, LOGOUT} from './mutation-types'
 Vue.use(Vuex)
 
 export default new Vuex.Store({
@@ -10,18 +11,18 @@ export default new Vuex.Store({
     user : {}
   },
   mutations: {
-    auth_request(state){
+    [AUTH_REQUEST](state){
       state.status = 'loading'
     },
-    auth_success(state, token, user){
+    [AUTH_SUCCESS](state, token, user){
       state.status = 'success'
       state.token = token
       state.user = user
     },
-    auth_error(state){
+    [AUTH_ERROR](state){
       state.status = 'error'
     },
-    logout(state){
+    [LOGOUT](state){
       state.status = ''
       state.token = ''
     },
